Pasan test de modificar la temperatura

+++ SimuladorGases/src/main/java/ar/edu/untref/simuladordegases/implementacion/Particula.java
@@ -17,7 +17,7 @@ public class Particula {
 
     protected Orientacion orientacion;
 
-	public Particula(Contenedor contenedor, float vel) {
+	public Particula(Contenedor contenedor, double vel) {
 
         this.contenedor = contenedor;
 		this.velocidad = vel;
